+++ R/session.R
@@ -321,8 +321,16 @@ html_table.rvest_session <- function(x,
                                header = NA,
                                trim = TRUE,
                                fill = deprecated(),
-                               dec = ".") {
-  html_table(read_html(x), header = header, trim = trim, fill = fill, dec = dec)
+                               dec = ".",
+                               na.strings = "NA") {
+  html_table(
+    read_html(x),
+    header = header,
+    trim = trim,
+    fill = fill,
+    dec = dec,
+    na.strings = na.strings
+  )
 }
 
 #' @export